Updated Project to 2018.2.20
Replaced Fizzyo Framework with the latest
Configured FizzyoProfile to current release - Note, this requires the GameID and Secret to be updated
Updated TextMeshPro
Replace use of Input.Getkey("Space") with FizzyoFramework.instance.device.ButtonPressed() - to get input from sensor or space

Signed-off-by: Simon (Darkside) Jackson <darkside@zenithmoon.com>

Only issue is I can't get the block to drop? unsure as to why.